+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.41.0 - TBD
+
+### Enhancements
+- Added three new methods to the Live client for more control over when I/O occurs. Together
+  they provide an alternative to `next_record()`:
+  - `fill_buf()`: reads available data from the socket into the internal buffer without
+    processing records
+  - `try_next_record()`: returns the next record from the internal buffer without
+     performing any I/O. Returns `Ok(None)` if no complete records are available in the buffer.
+  - `is_closed()`: returns whether the connection has been closed
 ## 0.40.0 - 2026-01-27
 
 ### Enhancements